Bug 519943 - Use CGFloat instead of float so that pixel scrolling works in 64 bit builds. r=josh
UltraBlame original commit: 675934009fc736f31bad81506a092ff5a4e09f8b

+++ widget/src/cocoa/nsChildView.h
@@ -105,8 +105,8 @@ extern "C" long TSMProcessRawKeyEvent(EventRef carbonEvent);
 
 
 @interface NSEvent (DeviceDelta)
-  - (float)deviceDeltaX;
-  - (float)deviceDeltaY;
+  - (CGFloat)deviceDeltaX;
+  - (CGFloat)deviceDeltaY;
 @end
 
 @interface ChildView : NSView<
